+++ anni/src/subcommands/mod.rs
@@ -12,4 +12,4 @@ pub use flac::FlacSubcommand;
 pub use library::LibrarySubcommand;
 pub use repo::RepoSubcommand;
 pub use split::SplitSubcommand;
-pub use workspace::{WorkspaceAction, WorkspaceSubcommand};
+pub use workspace::WorkspaceSubcommand;