+++ anni-workspace/CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.2.2
 
 - Make scan result of `AnniWorkspace::scan` a `BTreeMap` instead of a `HashMap`
 - Added `AnniWorkspace::new` to quickly find a workspace from `current_dir`
 - Added `AnniWorkspace::open` to open a workspace from a path without checking its parents recursively
+- Upgrade `anni-repo` to `0.4.1`
 
 ## 0.2.1